// src/pages/Game/GoldGame/types.d.ts
export type ScreenType = 'welcome' | 'playing' | 'summary'

export type ImageResourcesType = {
  bg: string
  bg_gold_item_1: string
  bg_gold_item_2: string
  btn: string
  btn_answer_off: string
  btn_back: string
  btn_null: string
  btn_sound_off: string
  btn_sound_on: string
  btn_summary: string
  game_name: string
  ico_board_correct: string
  ico_board_summary: string
  ico_board_text: string
  ico_diamond_box: string
  ico_gold: string
  ico_gold_box: string
  ico_heart_empty: string
  ico_heart_full: string
  ico_hook: string
  ico_man: string
  ico_result_correct: string
  ico_result_wrong: string
  ico_robo_happy: string
  ico_robo_shy: string
  ico_robo_sad: string
  ico_rope: string
  ico_scroll: string
  ico_sound: string
  ico_summary_logo: string
  ico_tnt: string
  logo: string
}
export type AudioResourcesType = {
  sound_correct: string
  sound_gold_correct: string
  sound_gold_correct_1: string
  sound_wrong: string
  theme: string
  theme_summary: string
}